+++ index.js
@@ -12,6 +12,7 @@ const mailchimp = require("@mailchimp/mailchimp_marketing")
 const { mong } = require('./db.js');
 const session = require("express-session")
 const passport = require("passport")
+const { getStartingPostion } = require ('./modules/starting-position.js')
 
 /////////////// Route links /////////////////////
 const compose = require('./routes/compose')
@@ -42,10 +43,41 @@ app.use(passport.session())
 
 // get requests
 // static pages
-app.get('/', (req, res) => {res.render("index")})
-app.get('/about', (req, res) => {res.render("about")})
-app.get('/services', (req, res) => {res.render("services")})
-app.get('/contact', (req, res) => {res.render("contact")})
+app.get('/', (req, res) => {
+  if(req.isAuthenticated()) {    
+  res.render("index")
+  } else {
+    console.log("not authorised as not logged in ");
+    res.redirect("/login")
+}
+})
+// app.get('/about', (req, res) => {res.render("about")})
+app.get('/about', (req, res) => {
+  if(req.isAuthenticated()) {    
+  res.render("about")
+  } else {
+    console.log("not authorised as not logged in ");
+    res.redirect("/login")
+}
+})
+// app.get('/services', (req, res) => {res.render("services")})
+app.get('/services', (req, res) => {
+  if(req.isAuthenticated()) {    
+  res.render("services")
+  } else {
+    console.log("not authorised as not logged in ");
+    res.redirect("/login")
+}
+})
+// app.get('/contact', (req, res) => {res.render("contact")})
+app.get('/contact', (req, res) => {
+  if(req.isAuthenticated()) {    
+  res.render("contact")
+  } else {
+    console.log("not authorised as not logged in ");
+    res.redirect("/login")
+}
+})
 
 app.get("/login", function(req, res){
   res.render("login");
@@ -117,7 +149,7 @@ app.get('/admin-articles/:articleName', function(req, res){
 })
 
 app.post("/login", passport.authenticate("local"), function(req, res){
-  res.redirect("/admin-articles");
+  res.redirect("/");
 });
 
 app.post("/register", function(req, res){
@@ -131,6 +163,63 @@ app.post("/register", function(req, res){
           })
       }
   })
+})
+// temporarily added in index js rather than app.use(articles....) whilst the site is protected
+app.get('/articles', function(req, res){
+  if(req.isAuthenticated()) {   
+    let headlineArticle = "61addf30f4c6b3b34388a9d5" 
+    articlesToShow = 3
+    articlesCurrentlyShown = articlesToShow
+   
+    let loadMoreClass = "load-more"
+    let loadMoreText = "Load More"
+  
+    Article.find({}, (err, foundItems) => {
+      articleArray = foundItems
+      totalArticles = articleArray.length
+      articlesLeftToShow = totalArticles - articlesCurrentlyShown
+      console.log(articlesLeftToShow);
+    res.render("articles", {
+        headlineArticle: headlineArticle,
+        articleArray: articleArray,
+        totalArticles: totalArticles,
+        startingArticleArrayPostion: getStartingPostion(foundItems, 0, articlesToShow),
+        endingArticleArrayPosition: foundItems.length - articlesToShow ,
+        href: "/articles/",
+        loadMoreClass: loadMoreClass,
+        loadMoreText: loadMoreText
+      })
+    })   
+  } else {
+      console.log("not logged in so not authorised to view admin page");
+      res.redirect("/login")
+  }
+})
+// temporarily added in index js rather than app.use(articles....) whilst the site is protected
+app.get('/articles/:articleName', function(req, res){
+  if(req.isAuthenticated()) {    
+    const typedTitle = _.kebabCase(_.lowerCase(req.params.articleName))
+
+    articleArray.forEach((post) => {
+        const storedTitle = _.kebabCase(_.lowerCase(post.title))
+      if(typedTitle ===  storedTitle) {
+  
+        const articleContent = post.content
+        const reformatedContent = articleContent.replace(/(\r\n|\r|\n)/g, '<br>') //converts \r\n text from the DB to <br> tags
+        const markedContent = marked.parse(articleContent)
+        res.render("article-page", {
+          title: post.title,
+          content: markedContent,
+          date: post.date,
+          imageLink: post.image,
+          altImage: post.imageAlt
+        })
+      }
+    })
+  } else {
+      console.log("not logged in so not authorised to view article admin page");
+      res.redirect("/login")
+  }
 })
 
 app.post("/failure", (req, res) => {
